# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QUEPON GGX Command System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /queue:
    get:
      operationId: listQueueEntries
      tags: [queue]
      summary: List all active queue entries (admin)
      responses:
        "200":
          description: Queue entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueueEntry"

    post:
      operationId: joinQueue
      tags: [queue]
      summary: Join the queue (player)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueueInput"
      responses:
        "201":
          description: Joined queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
        "400":
          description: Already in queue or validation error
  /queue/my:
    get:
      operationId: getMyQueueEntry
      tags: [queue]
      summary: Get the current player's active queue entry
      responses:
        "200":
          description: My queue entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
        "404":
          description: Not in queue
  /queue/{queueId}/approve:
    patch:
      operationId: approveQueueEntry
      tags: [queue]
      summary: Approve a queue entry (admin)
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
  /queue/{queueId}/remove:
    patch:
      operationId: removeQueueEntry
      tags: [queue]
      summary: Remove/cancel a queue entry (admin or player self)
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
  /queue/{queueId}/assign:
    patch:
      operationId: assignQueueEntryToPc
      tags: [queue]
      summary: Assign queue entry to a PC and start session (admin)
      parameters:
        - name: queueId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignPcInput"
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"

  # --- SESSIONS ---
...
components:
  schemas:
...
    QueueEntry:
      type: object
      required: [id, userId, username, status, position, joinedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        status:
          type: string
          enum: [waitingApproval, approved, waiting, assigned, cancelled, removed, noShow, completed]
        position:
          type: integer
        requestedTier:
          type: ["string", "null"]
        preferredPcId:
          type: ["string", "null"]
        estimatedWaitMinutes:
          type: ["integer", "null"]
        approvedBy:
          type: ["string", "null"]
        assignedPcId:
          type: ["string", "null"]
        sessionId:
          type: ["string", "null"]
        joinedAt:
          type: string
        approvedAt:
          type: ["string", "null"]
        assignedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    QueueInput:
      type: object
      properties:
        requestedTier:
          type: string
        preferredPcId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    AssignPcInput:
      type: object
      required: [pcId]
      properties:
        pcId:
          type: string
        durationMinutes:
          type: integer

    # --- SESSIONS ---
    Session:
      type: object
      required: [id, userId, username, pcId, status, ratePerHour, durationMinutes, startedAt, endsAt]
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        pcId:
          type: string
        pcLabel:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, locked, extended, completed, cancelled, abandoned]
        ratePerHour:
          type: number
        durationMinutes:
          type: integer
        extendedMinutes:
          type: integer
        startedAt:
          type: string
        endsAt:
          type: string
        endedAt:
          type: ["string", "null"]
        remainingSeconds:
          type: ["integer", "null"]
        costSoFar:
          type: number
        finalCost:
          type: ["number", "null"]
        paymentSource:
          type: string
        queueId:
          type: ["string", "null"]
        isLocked:
          type: boolean
        createdAt:
          type: string
...